Add a recolve_name_mut method to LuaState
This is going to be useful for when we need to mutate local variables
and stuff, typically in assignment statements.

We need both variants as the first resolv_name is going to be for
evaluation purposes, which in no way should be able to mutate scopes!

NB: the Lua spec actually tells us that evaluating something, even when
it's not a function call, can do anything, including, of course, random
mutations of the scopes. But since we don't have any functions yet we're
not there just yet.